// src/modules/resources.generated.ts
export type ResourceAction = "add" | "del" | "delete" | "download" | "edit" | "find" | "findmain" | "fiscalize" | "get" | "notes" | "send" | "trigger" | "unfiscalize";

export type ResourceName = "companies" | "company_accounts" | "company_addresses" | "company_packs" | "contractors" | "declaration_body_jpkvat" | "declaration_body_pit" | "declaration_countries" | "documents" | "expenses" | "goods" | "interests" | "invoice_deliveries" | "invoice_descriptions" | "invoices" | "ledger_accountant_years" | "ledger_operation_schemas" | "notes" | "payment_cashboxes" | "payments" | "series" | "tags" | "taxregisters" | "term_groups" | "terms" | "translation_languages" | "user_companies" | "users" | "vat_codes" | "vehicle_run_rates" | "vehicles" | "warehouse_document_p_w" | "warehouse_document_p_z" | "warehouse_document_r" | "warehouse_document_r_w" | "warehouse_document_w_z" | "warehouse_document_z_d" | "warehouse_document_z_p_d" | "warehouse_document_z_p_m" | "warehouses" | "webhooks";

export type ActionDef = {
  method: string;
  wrapPayload?: boolean;
};

export type ResourceDefinition = {
  name: ResourceName;
  singularName: string;
  pluralName: string;
  actions: Record<string, ActionDef>;
};